+++ source/extensions/filters/network/redis_proxy/command_splitter_impl.cc
@@ -197,10 +197,10 @@ SplitRequestPtr MGETRequest::create(Router& router, Common::Redis::RespValuePtr&
     const auto route = router.upstreamPool(base_request->asArray()[i].asString());
     if (route) {
       // Create composite array for a single get.
-      Common::Redis::RespValue single_get(base_request,
-                                          Common::Redis::Utility::GetRequest::instance(), i, i);
+      const Common::Redis::RespValue single_mget(
+          base_request, Common::Redis::Utility::GetRequest::instance(), i, i);
       pending_request.handle_ = makeFragmentedRequest(
-          route, "get", base_request->asArray()[i].asString(), single_get, pending_request);
+          route, "get", base_request->asArray()[i].asString(), single_mget, pending_request);
     }
 
     if (!pending_request.handle_) {
@@ -266,16 +266,16 @@ SplitRequestPtr MSETRequest::create(Router& router, Common::Redis::RespValuePtr&
   request_ptr->pending_response_->type(Common::Redis::RespType::SimpleString);
 
   Common::Redis::RespValueSharedPtr base_request = std::move(incoming_request);
-  uint64_t fragment_index = 0;
-  for (uint64_t i = 1; i < base_request->asArray().size(); i += 2) {
+  unsigned fragment_index = 0;
+  for (unsigned i = 1; i < base_request->asArray().size(); i += 2) {
     request_ptr->pending_requests_.emplace_back(*request_ptr, fragment_index++);
     PendingRequest& pending_request = request_ptr->pending_requests_.back();
 
     const auto route = router.upstreamPool(base_request->asArray()[i].asString());
     if (route) {
       // Create composite array for a single set command.
-      Common::Redis::RespValue single_set(base_request,
-                                          Common::Redis::Utility::SetRequest::instance(), i, i + 1);
+      const Common::Redis::RespValue single_set(
+          base_request, Common::Redis::Utility::SetRequest::instance(), i, i + 1);
       ENVOY_LOG(debug, "redis: parallel set: '{}'", single_set.toString());
       pending_request.handle_ = makeFragmentedRequest(
           route, "set", base_request->asArray()[i].asString(), single_set, pending_request);
@@ -337,12 +337,12 @@ SplitRequestPtr SplitKeysSumResultRequest::create(Router& router,
   request_ptr->pending_response_->type(Common::Redis::RespType::Integer);
 
   Common::Redis::RespValueSharedPtr base_request = std::move(incoming_request);
-  for (uint64_t i = 1; i < base_request->asArray().size(); i++) {
+  for (unsigned i = 1; i < base_request->asArray().size(); i++) {
     request_ptr->pending_requests_.emplace_back(*request_ptr, i - 1);
     PendingRequest& pending_request = request_ptr->pending_requests_.back();
 
     // Create the composite array for a single fragment.
-    Common::Redis::RespValue single_fragment(base_request, base_request->asArray()[0], i, i);
+    const Common::Redis::RespValue single_fragment(base_request, base_request->asArray()[0], i, i);
     ENVOY_LOG(debug, "redis: parallel {}: '{}'", base_request->asArray()[0].asString(),
               single_fragment.toString());
     const auto route = router.upstreamPool(base_request->asArray()[i].asString());