+++ source/extensions/filters/network/redis_proxy/conn_pool_impl.h
@@ -66,15 +66,13 @@ public:
                                                   PoolCallbacks& callbacks) override;
   /**
    * Makes a redis request based on IP address and TCP port of the upstream host (e.g.,
-   moved/ask cluster redirection). This is now only kept mostly for testing.
+   * moved/ask cluster redirection). This is now only kept mostly for testing.
    * @param host_address supplies the IP address and TCP port of the upstream host to receive
-   the
-   * request.
+   * the request.
    * @param request supplies the Redis request to make.
    * @param callbacks supplies the request completion callbacks.
    * @return PoolRequest* a handle to the active request or nullptr if the request could not be
-   made
-   *         for some reason.
+   * made for some reason.
    */
   Common::Redis::Client::PoolRequest*
   makeRequestToHost(const std::string& host_address, const Common::Redis::RespValue& request,